+++ pkg/bindings/connection.go
@@ -1,6 +1,7 @@
 package bindings
 
 import (
+	"bytes"
 	"context"
 	"errors"
 	"fmt"
@@ -9,6 +10,7 @@ import (
 	"net/http"
 	"net/url"
 	"os"
+	"os/user"
 	"strconv"
 	"strings"
 	"time"
@@ -16,6 +18,7 @@ import (
 	"github.com/blang/semver/v4"
 	"github.com/containers/common/pkg/ssh"
 	"github.com/containers/podman/v5/version"
+	"github.com/kevinburke/ssh_config"
 	"github.com/sirupsen/logrus"
 	"golang.org/x/net/proxy"
 )
@@ -149,6 +152,14 @@ func sshClient(_url *url.URL, uri string, identity string, machine bool) (Connec
 	connection := Connection{
 		URI: _url,
 	}
+	userinfo := _url.User
+	if _url.User == nil {
+		u, err := user.Current()
+		if err != nil {
+			return connection, fmt.Errorf("current user could not be determined: %w", err)
+		}
+		userinfo = url.User(u.Username)
+	}
 	port := 22
 	if _url.Port() != "" {
 		port, err = strconv.Atoi(_url.Port())
@@ -156,15 +167,65 @@ func sshClient(_url *url.URL, uri string, identity string, machine bool) (Connec
 			return connection, err
 		}
 	}
+	// ssh_config
+	alias := _url.Hostname()
+	cfg := ssh_config.DefaultUserSettings
+	found := false
+	if val := cfg.Get(alias, "User"); val != "" {
+		userinfo = url.User(val)
+		found = true
+	}
+	if val := cfg.Get(alias, "Hostname"); val != "" {
+		uri = val
+		found = true
+	}
+	if val := cfg.Get(alias, "Port"); val != "" {
+		if val != ssh_config.Default("Port") {
+			port, err = strconv.Atoi(val)
+			if err != nil {
+				return connection, fmt.Errorf("port is not an int: %s: %w", val, err)
+			}
+			found = true
+		}
+	}
+	if val := cfg.Get(alias, "IdentityFile"); val != "" {
+		if val != ssh_config.Default("IdentityFile") {
+			identity = strings.Trim(val, "\"")
+			found = true
+		}
+	}
+	if found {
+		logrus.Debugf("ssh_config alias found: %s", alias)
+		logrus.Debugf("  User: %s", userinfo.Username())
+		logrus.Debugf("  Hostname: %s", uri)
+		logrus.Debugf("  Port: %d", port)
+		logrus.Debugf("  IdentityFile: %q", identity)
+	}
 	conn, err := ssh.Dial(&ssh.ConnectionDialOptions{
 		Host:                        uri,
 		Identity:                    identity,
-		User:                        _url.User,
+		User:                        userinfo,
 		Port:                        port,
 		InsecureIsMachineConnection: machine,
 	}, ssh.GolangMode)
 	if err != nil {
 		return connection, newConnectError(err)
+	}
+	if _url.Path == "" {
+		session, err := conn.NewSession()
+		if err != nil {
+			return connection, err
+		}
+		defer session.Close()
+
+		var b bytes.Buffer
+		session.Stdout = &b
+		if err := session.Run(
+			"podman info --format '{{.Host.RemoteSocket.Path}}'"); err != nil {
+			return connection, err
+		}
+		val := strings.TrimSuffix(b.String(), "\n")
+		_url.Path = val
 	}
 	dialContext := func(ctx context.Context, _, _ string) (net.Conn, error) {
 		return ssh.DialNet(conn, "unix", _url)